services/crm-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
requests==2.31.0
httpx==0.25.2
python-multipart==0.0.6
slowapi==0.1.9
